+++ src/Convo/Core/Adapters/Alexa/AmazonAuthService.php
@@ -115,8 +115,16 @@ class AmazonAuthService
 
         $expires = $amazonConfig['expires_in'];
         $created = $amazonConfig['created'] ?? 0;
-
-        if ($now < $created + $expires) {
+        
+        $this->_logger->debug('Got client_auth timestamps created['.$created.'] expires['.$expires.'] now['.$now.']');
+        
+        if ( isset( $amazonConfig['token_type']) && $amazonConfig['token_type'] === 'bearer') {
+            $expired    =   $now < $created + $expires;
+        } else {
+            $expired    =   $now < $expires;
+        }
+        
+        if ( $expired) {
             $this->_logger->debug('No need to refresh token yet.');
             return;
         }
@@ -157,15 +165,15 @@ class AmazonAuthService
         }
     }
 
-    public function storeAuthCredentials(IAdminUser $user, $credentials)
+    public function storeAuthCredentials( IAdminUser $user, $credentials)
     {
         $config = [
             'amazon' => [
                 'client_auth' => $credentials
             ]
         ];
-
-        $this->_adminUserDataProvider->updatePlatformConfig($user->getId(), $config);
+        
+        $this->_adminUserDataProvider->updatePlatformConfig( $user->getId(), $config);
     }
 
     public function getAuthCredentials(IAdminUser $user)